+++ app/cronjob/trailingTrade/step/handle-open-orders.js
@@ -24,8 +24,8 @@ const execute = async (logger, rawData) => {
     action,
     isLocked,
     openOrders,
-    buy: { limitPrice: buyLimitPrice },
-    sell: { limitPrice: sellLimitPrice }
+    buy: { limitPrice: buyLimitPrice, currentPrice: buyCurrentPrice },
+    sell: { limitPrice: sellLimitPrice, currentPrice: sellCurrentPrice }
   } = data;
 
   if (isLocked) {
@@ -83,12 +83,25 @@ const execute = async (logger, rawData) => {
 
           data.accountInfo = await getAccountInfoFromAPI(logger);
         }
-      } else if (parseFloat(order.stopPrice) >= buyLimitPrice) {
-        // Is the stop price is higher than current limit price?
-        logger.info(
-          { stopPrice: order.stopPrice, buyLimitPrice, saveLog: true },
-          'Stop price is higher than buy limit price, cancel current buy order'
-        );
+      } else if (
+        parseFloat(order.stopPrice) > buyLimitPrice ||
+        buyCurrentPrice > parseFloat(order.price)
+      ) {
+        if (parseFloat(order.stopPrice) > buyLimitPrice) {
+          // Is the buy order stop price higher than current buy limit price?
+          logger.info(
+            { stopPrice: order.stopPrice, buyLimitPrice, saveLog: true },
+            'Buy order stop price is higher than current buy limit price, cancel current buy order'
+          );
+        }
+
+        if (buyCurrentPrice > parseFloat(order.price)) {
+          // Is the current price higher than buy order price?
+          logger.info(
+            { stopPrice: order.stopPrice, buyLimitPrice, saveLog: true },
+            'Current price is higher than buy order price, cancel current buy order'
+          );
+        }
 
         // Cancel current order
         const cancelResult = await cancelOrder(logger, symbol, order);
@@ -140,12 +153,25 @@ const execute = async (logger, rawData) => {
     }
 
     if (order.side.toLowerCase() === 'sell') {
-      // Is the stop price is less than current limit price?
-      if (parseFloat(order.stopPrice) <= sellLimitPrice) {
-        logger.info(
-          { stopPrice: order.stopPrice, sellLimitPrice, saveLog: true },
-          'Stop price is less than sell limit price, cancel current sell order'
-        );
+      if (
+        parseFloat(order.stopPrice) < sellLimitPrice ||
+        sellCurrentPrice < parseFloat(order.price)
+      ) {
+        if (parseFloat(order.stopPrice) < sellLimitPrice) {
+          // Is the sell order stop price lower than current sell limit price?
+          logger.info(
+            { stopPrice: order.stopPrice, sellLimitPrice, saveLog: true },
+            'Sell order stop price is less than current sell limit price, cancel current sell order'
+          );
+        }
+
+        if (sellCurrentPrice < parseFloat(order.price)) {
+          // Is the current price less than sell order price?
+          logger.info(
+            { stopPrice: order.stopPrice, sellLimitPrice, saveLog: true },
+            'Current price is less than sell order price, cancel current sell order'
+          );
+        }
 
         // Cancel current order
         const cancelResult = await cancelOrder(logger, symbol, order);